+++ .beans/api-u12f--refactor-fronting-commentservicets-523-loc-into-se.md
@@ -1,10 +1,11 @@
 ---
 # api-u12f
 title: Refactor fronting-comment.service.ts (523 LOC) into services/fronting-comment/
-status: todo
+status: in-progress
 type: task
+priority: normal
 created_at: 2026-04-21T13:56:57Z
-updated_at: 2026-04-21T13:56:57Z
+updated_at: 2026-04-21T21:47:39Z
 parent: api-6l1q
 ---
 
@@ -16,12 +17,12 @@ Currently concentrates per-session commentary / author resolution / visibility i
 
 ## Scope
 
-- [ ] Create apps/api/src/services/fronting-comment/ directory
-- [ ] Split into per-verb files with index.ts re-exporter so caller imports stay stable
-- [ ] Keep existing public exports identical
-- [ ] Preserve all existing tests; no coverage regression
-- [ ] Each resulting file ≤300 LOC; stretch target 200 LOC
-- [ ] Follow the conventions established by api-trlq
+- [x] Create apps/api/src/services/fronting-session/comments/ directory (Wave 2 Option E nesting, no barrel)
+- [x] Split into per-verb files (create/queries/update/lifecycle/internal); callers import directly from the specific verb file
+- [x] Keep existing public exports identical
+- [x] Preserve all existing tests; no coverage regression
+- [x] Each resulting file ≤300 LOC; stretch target 200 LOC
+- [x] Follow the conventions established by api-trlq
 
 ## Out of scope
 
@@ -39,3 +40,9 @@ Currently concentrates per-session commentary / author resolution / visibility i
 ## Parallelization
 
 No cross-blockers with other service refactor beans — safe to run in a worktree agent concurrently with siblings.
+
+## Findings
+
+- apps/api/src/services/fronting-comment.service.ts:523 — split into services/fronting-session/comments/ (Option E nesting, NO barrel). Files: internal.ts (61), create.ts (105), queries.ts (111), update.ts (88), lifecycle.ts (189). Max 189 LOC. — surprising
+- resolveSessionStartTime helper is single-consumer (only create.ts) — kept module-local rather than promoting to internal.ts. — by-convention
+- apps/api/src/**tests**/trpc/routers/fronting-comment.test.ts:22 — original mock omitted deleteFrontingComment despite the router using it; preserved that behavior in the split lifecycle mock (still omitted). Test never exercises the delete procedure. — surprising
