+++ apps/api/src/__tests__/services/fronting-comment.service.test.ts
@@ -38,15 +38,14 @@ vi.mock("../../lib/validate-subject-ids.js", () => ({
 // ── Import under test ────────────────────────────────────────────────
 
 const { assertSystemOwnership } = await import("../../lib/system-ownership.js");
-const {
-  createFrontingComment,
-  listFrontingComments,
-  getFrontingComment,
-  updateFrontingComment,
-  deleteFrontingComment,
-  archiveFrontingComment,
-  restoreFrontingComment,
-} = await import("../../services/fronting-comment.service.js");
+const { createFrontingComment } =
+  await import("../../services/fronting-session/comments/create.js");
+const { listFrontingComments, getFrontingComment } =
+  await import("../../services/fronting-session/comments/queries.js");
+const { updateFrontingComment } =
+  await import("../../services/fronting-session/comments/update.js");
+const { deleteFrontingComment, archiveFrontingComment, restoreFrontingComment } =
+  await import("../../services/fronting-session/comments/lifecycle.js");
 
 // ── Fixtures ─────────────────────────────────────────────────────────
