+++ apps/api/src/routes/fronting-sessions/comments/create.ts
@@ -9,7 +9,7 @@ import { parseJsonBody } from "../../../lib/parse-json-body.js";
 import { envelope } from "../../../lib/response.js";
 import { createIdempotencyMiddleware } from "../../../middleware/idempotency.js";
 import { createCategoryRateLimiter } from "../../../middleware/rate-limit.js";
-import { createFrontingComment } from "../../../services/fronting-comment.service.js";
+import { createFrontingComment } from "../../../services/fronting-session/comments/create.js";
 
 import type { AuthEnv } from "../../../lib/auth-context.js";
