+++ apps/api/src/routes/fronting-sessions/comments/get.ts
@@ -5,7 +5,7 @@ import { getDb } from "../../../lib/db.js";
 import { parseIdParam, requireIdParam } from "../../../lib/id-param.js";
 import { envelope } from "../../../lib/response.js";
 import { createCategoryRateLimiter } from "../../../middleware/rate-limit.js";
-import { getFrontingComment } from "../../../services/fronting-comment.service.js";
+import { getFrontingComment } from "../../../services/fronting-session/comments/queries.js";
 
 import type { AuthEnv } from "../../../lib/auth-context.js";
